move frame and back button to page components
