+++ web-ui/src/pages/add.tsx
@@ -1,9 +1,9 @@
 import styles from './add.less';
-import Frame from '@/components/main/frame';
+import Frame from '@/components/page/frame/frame';
 import { registerPageLoad } from '@/matomo';
 import { Col, Row } from 'antd';
 import { useEffect, useState } from 'react';
-import BackButton from '@/components/main/back/back';
+import BackButton from '@/components/page/back/back';
 import AddLecture from '@/components/input/add-lecture/add-lecture';
 
 export default function IndexPage() {
